requirements.txt
appdirs==1.4.0
click==6.7
Flask==0.12
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.9.5
jsonrpc2==0.4.1
MarkupSafe==0.23
packaging==16.8
psycopg2==2.6.2
pyparsing==2.1.10
python-dateutil==2.6.0
PyYAML==3.12
six==1.10.0
SQLAlchemy==1.1.5
Werkzeug==0.11.15
